feat (product-service) : add product routes

+++ product/src/useCase/deleteProductUseCase.ts
@@ -2,6 +2,8 @@ import { ProductRepoType } from "../database/index.js";
 
 export default function buildDeleteProduct(productRepository: ProductRepoType) {
   return async function deleteProduct({ id }: { id: string }) {
+    if(!id) throw new Error("Product id is missing");
+
     return await productRepository.delete(id);
   };
 }
